requirements.txt
asn1crypto==0.24.0
Authlib==0.12.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.6
WTForms==2.2.1
